news/articles/functions/news_articles_ingest_get_paper/requirements.txt
google-cloud-secret-manager==2.2.0
google-cloud-firestore==2.0.2
google-cloud-pubsub==1.5.0
elasticsearch==7.13.4
beautifulsoup4>=4.4.1
cssselect>=0.9.2
feedfinder2>=0.0.4
feedparser>=5.2.1
jieba3k>=0.35.1
lxml>=3.6.0
nltk>=3.2.1
Pillow>=3.3.0
pythainlp>=1.7.2
python-dateutil>=2.5.3
PyYAML>=3.11
requests>=2.10.0
tinysegmenter==0.3
tldextract>=2.0.1
